Addressing: Rename apps/Llama2App.py to follow snake case convention #665
